// src/app/models/EventData.ts
export interface EventData {
  date: string
  time: string
  icon: string
  Name: string
  genre: string
  venue: string
  eventId: string;
  segments: string;
  subGenre: string;
  type: string;
  artists: string[];
  ticketStatus: string;
  seatMap: string;
  buyTickets: string;
  priceRanges: string;
}